// src/models/common/Hardpoint.ts
export enum Hardpoint {
  Empty = 'Empty',
  Sensors = 'Sensors',
  Shoulder = 'Shoulder',
  UpperArmActuator = 'UpperArmActuator',
  LowerArmActuator = 'LowerArmActuator',
  Hand = 'Hand',
  Hip = 'Hip',
  UpperLegActuator = 'UpperLegActuator',
  LowerLegActuator = 'LowerLegActuator',
  FootActuator = 'FootActuator',
  Cockpit = 'Cockpit',
  LifeSupport = 'LifeSupport',
  Engine = 'Engine',
  Gyro = 'Gyro',
  Energy = 'Energy',
  Ballistic = 'Ballistic',
  Missile = 'Missile',
  Omni = 'Omni',
  MissileAmmo = 'MissileAmmo',
  BallisticAmmo = 'BallisticAmmo',
  EnergyAmmo = 'EnergyAmmo',
  HeatSink = 'HeatSink',
  JumpJet = 'JumpJet',
  InternalStructure = 'InternalStructure',
  Armor = 'Armor',
}